# backend-go/api-spec.yml
openapi: 3.0.3
info:
  title: Email Helper API
  description: REST API for Email Helper application with AI-powered email classification and task management
  version: 1.0.0
  contact:
    name: Email Helper
servers:
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /api/emails/extract-tasks:
    post:
      summary: Extract tasks from emails
      tags: [Emails]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email_ids
              properties:
                email_ids:
                  type: array
                  items:
                    type: string
                apply_to_outlook:
                  type: boolean
      responses:
        '200':
          description: Task extraction completed
  /api/emails/sync-to-database:
    post:
      summary: Sync classified emails to database
      tags: [Emails]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - emails
              properties:
                emails:
                  type: array
                  items:
                    $ref: '#/components/schemas/Email'
      responses:
        '200':
          description: Sync completed
...
  /api/processing/start:
    post:
      summary: Start processing pipeline
      tags: [Processing]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email_ids
              properties:
                email_ids:
                  type: array
                  items:
                    type: string
                priority:
                  type: string
                  enum: [low, medium, high, urgent]
                  default: medium
      responses:
        '200':
          description: Pipeline started
          content:
            application/json:
              schema:
                type: object
                properties:
                  pipeline_id:
                    type: string
                  status:
                    type: string
                  email_count:
                    type: integer
                  message:
                    type: string
  /api/processing/{id}/status:
    get:
      summary: Get pipeline status
      tags: [Processing]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Pipeline status
        '404':
          description: Pipeline not found
components:
  schemas:
    Email:
      type: object
      properties:
        id:
          type: string
        subject:
          type: string
        sender:
          type: string
        recipient:
          type: string
        body:
          type: string
        content:
          type: string
        received_time:
          type: string
          format: date-time
        date:
          type: string
          format: date-time
        is_read:
          type: boolean
        has_attachments:
          type: boolean
        importance:
          type: string
          enum: [Low, Normal, High]
        categories:
          type: array
          items:
            type: string
        conversation_id:
          type: string
        conversation_count:
          type: integer
        ai_category:
          type: string
        ai_confidence:
          type: number
        ai_reasoning:
          type: string
        one_line_summary:
          type: string
        category:
          type: string
...